Support tracking of containing contexts
In this commit, we now retrieve the Clang semantic parent as an analogue
of Cscope's containing function. I don't know if we ever need this
except for the line-oriented interface, but it's nice to have it anyway.
Note that the results of this differ significantly from what Cscope
considers to be the "parent" of a symbol, but in a seemingly useful way.