show containing function when finding definitions in the curses UI
It turns out, while not relevant in Cscope, this is relevant in Clink because
our search is more broad. It is finding *any* definition, not just global ones.
In a world of C++, this seems like a more useful search to me, but only
experience will tell. Note that the columns are unified now (containing function
also makes sense for #includes in code bases insane enough to include files
within a function) so we're on our way to removing this information from result
rows altogether.