restructure Vim highlighter as a standalone function
Looking forwards, we only ever want to read and highlight the entire contents of
a file, not selective lines that we'll pull out one-by-one. With this in mind,
there's no point bouncing back and forth with calls to VimHighlighter::get_line.
Instead, we just return a vector of all the lines from the file.