restructure database to normalise a bit
The idea behind this change is to separate content from symbol information so
that we can read files out of band and dump them into the content table. The
latter is not yet done. Note that this commit also fixes a memory leak wherein
Database::m_delete was not being cleaned up.