libclink: return an error code from Database::add instead of exception
It turns out we can get irrelevant failures here because the C/C++ parser yields
symbols from more than just the file it was asked to parse. I.e. it may end up
giving us symbols that we have already added to the database at some previous
point in the update.

In the future, perhaps we should prune these out given -- with a proper queueing
and file discovery algorithm -- we would discover and parse these files
naturally on their own.