fix: bind SQL text parameters to something with long enough lifetime
We are binding these parameters with SQLITE_STATIC, which tells SQLite that it
can assume the lifetime of the string parameter is long enough to take a const
pointer to it throughout the query lifetime. This is not true with the current
iterator-driven design. Or more accurately, we were using the wrong pointer. We
now use pointers from the iterator state that are guaranteed to be valid until
the iterator itself is destroyed.