allow missing context in line-oriented interface
This avoids an inconsistent state where we believe we have the context of a file
but have actually parsed but not syntax highlighted the file. E.g. the user
Ctrl+C-ed Clink while it was building the database. Rather than trying to
enforce that the context is always there, it is simpler to just drop this as an
assumption altogether.