remove Cscope legacy command line options
It seems more straightforward to implement Clink itself without these
constraints, defining its own options independent of Cscope precedent. If
someone needs a drop-in replacement for the `cscope` binary, this can be
implemented as a wrapper script that remaps Cscope options to their Clink
equivalents.

For the same reason, I think we should remove the line-oriented interface and
implement this as a separate binary. This would probably result in more
flexibility for both – the main Clink binary no longer needs to care about
“mode” and the line-oriented interface can assume it has no UI.