suppress deleted files in results
Files that are not found no longer appear in results in the NCurses interface.
This does not affect the line-oriented interface, but the plan is to refactor
that into a separate binary soon.

It seemed cleaner to do this suppression during result retrieval rather than
either of the obvious alternatives:

  1. Suppress results for deleted files in libclink during retrieval. This would
     make it impossible to find such dangling results using the libclink API.
     The clink binary does not do this, but some future tool might.

  2. Remove database records for deleted files during the build phase. This
     seems like it would incur a relatively expensive operation (enumerating all
     database records and confirming they match existing on-disk files) for a
     relatively rare event. Well maybe not so rare, but at least easily
     handleable elsewhere using a cheaper strategy.

(2) would also suffer from a TOCTOU issue where we could still encountered
deleted files anyway. The solution we landed on also suffers from this but at
least the window is smaller and more intuitive.

Closes #17 “suppress results for deleted files”