add an API function for versioning
This adds a versioning mechanism based on that from Rumur¹ that has served us
well for years. This incurs a Python ≥3.6 dependency, but it is assumed this is
not a problem for anyone who already has libclang available. This side steps a
couple of issues with the Rumur implementation of this. In particular, the
version is generated in a source file rather than a header to avoid:

  1. Every build of a new Git commit updating version.h causing a complete
     rebuild of anything including this file; and

  2. Versioning being available in a header confusing client programs that have
     an inconsistent include vs link path, causing them to think they have one
     version of the library when they are really linking against another.

¹ https://github.com/smattr/rumur