run: split the PTY, giving the child the subordinate end
The prior code was giving the child the controlling end of the PTY. Technically
there is nothing wrong with this, but it is somewhat backwards. This change is a
steps towards solving a further problem, that we are not connecting the other
end of the PTY to anything. This appears to be why this technique results in a
deadlock on macOS.

Github: related to #6 “slow Vim cat on non-Linux”