run: fix: use PTY strategy on macOS as well
Following the prior changes, the PTY strategy works (does not cause a deadlock)
on macOS. This means we can switch to it, avoiding a 2 second delay when
vim-catting or vim-highlighting. This is a fairly massive win (vim-highlighting
in a large source tree was easily the bottleneck during build).

Github: fixes #6 “slow Vim cat on non-Linux”