call printf instead of echo in run() tests
On macOS, `echo` seems to perform an `fflush` that fails when using a PTY, as we
will in upcoming changes. This is somewhat irrelevant to us (we do not care
lost flushes in this scenario and most programs will not even check the return
value of `fflush`). POSIX `printf` does not seem to have the same issue.

Github: related to #6 “slow Vim cat on non-Linux”