fix: tell Vim to skip TTY checks on stdin/stdout
The `--not-a-term` option is undocumented, but inspecting the Vim source shows
it controls a call to `check_tty`, which turns out to be the exact cause of our
slowdown. By default, Vim performs this check and stalls for 2005 milliseconds
if either stdin or stdout are not TTYs.

This should also allow us to have a unified stdout masking strategy on all
operating systems, but this is left for an upcoming commit.

Github: fixes #6 “slow Vim cat on non-Linux”