remove PTY strategy of masking stdout on non-Linux
As of the prior commit, the /dev/null strategy is fast enough to use everywhere.
I have not rigorously profiled, but I would expect it to be *faster* than the
PTY strategy, simply because it is doing something less complicated. That is,
the combined effect of this commit and the prior one should be a speed up on all
platforms.

Github: related to #6 “slow Vim cat on non-Linux”