search by suffix stem for #includes, rather than full path
When searching for a #include of something like "clink.h" in the Clink code
base, no results would be found because all #includes of it are as
"clink/clink.h". The clink_db_find_includer function was looking for an exact
match. This behaviour differs from Cscope and is counter intuitive to the user.

Searches like this are now done as a suffix. If your search term is "clink.h"
the results include everything ending in clink.h.

The implementation in this commit is suboptimal, in that '%' and '_' characters
in the search term will be interpreted as wildcards. We should improve this in
future.