use file size as a hash
This is not necessarily a good choice, but it is better than the previous
behaviour of giving every file the hash 0. It seems reasonable to assume that a
file that has changed has a modified time or file size. If a file has changed,
but still has both these values the same, the user is probably doing something
quirky and they know it.