split Clang visitor prelude into a thin wrapper
The motivation for this is progressing to an alternative C/C++ parsing function
that pumps results directly into the database. The existing parsing via a manual
iterator is too slow and, while I do not know that the iteration patterns are a
meaningful bottleneck here, they are certainly unnecessary overhead right now.

Future commits will move to have two separate visitor functions with a common
core.