move cursor enqueue to the end of the Clang visitor
The motivation for this is progressing to an alternative C/C++ parsing function
that pumps results directly into the database. The existing parsing via a manual
iterator is too slow and, while I do not know that the iteration patterns are a
meaningful bottleneck here, they are certainly unnecessary overhead right now.

This swaps the order of enqueue of the cursor and the symbol-to-be-yielded. This
has no immediate relevant effect, but the alternative visitor described above
will need to handle enqueuing – or more accurately, descending recursively – at
the end of the visitor.