parametrise Clang visitor with yield and enqueue functions
The motivation for this is progressing to an alternative C/C++ parsing function
that pumps results directly into the database. The existing parsing via a manual
iterator is too slow and, while I do not know that the iteration patterns are a
meaningful bottleneck here, they are certainly unnecessary overhead right now.

This change alters the core of the Clang visitor to, instead of calling
`push_symbol` and `push_cursor` directly, to call function pointers passed in as
parameters. The only use of this currently is with `push_symbol` and
`push_cursor` passed in as these parameters, but the alternative visitor
described above will pass in different functions for these.

Note that this all required altering `push_cursor` and `push_symbol` to write to
the `rc` member of the state instead of relying on their caller to do this. To
see why this was necessary, observe that the call sites of these functions
(through the `yield` and `enqueue` pointers) no longer knows the type of the
state.