rename C parsers iterator state to iter_state_t
The motivation for this is progressing to an alternative C/C++ parsing function
that pumps results directly into the database. The existing parsing via a manual
iterator is too slow and, while I do not know that the iteration patterns are a
meaningful bottleneck here, they are certainly unnecessary overhead right now.

An upcoming change will introduce the above described one shot parser. It will
have its own state. To prepare for this, we specialise the name of the state
type for the iterative parser.