add an API for one-shot parsing of C/C++
The overhead of constructing and managing an iterator when parsing C/C++ is an
unnecessary burden in the clink binary. Its iteration is straightforward and all
it wants to do is insert each result into a database. Providing a way to avoid
this overhead, the present commit exposes a function for directly inserting the
parsed results into a database.