implement a new Vim highlighting interface
Some high level sampling of the existing Vim highlighting slowness suggests the
bottleneck is actually Vim’s TOhtml.vim. This translator is written in Vimscript
and is unlikely to have been intended for high throughput. So rather than
attempting to optimise what appears to be an inherently limited design, this
commit implements a different strategy for highlighting. This function fakes
terminal dimensions to ensure a file is rendered as-is and then asks Vim to
render it to a pipe.

Some coarse-grained profiling indicates this is ~99.2% faster than
`vim_highlight`.

This is not yet wired up to either `vim-cat` or `clink`’s highlighter. When this
further step has been done, we should be able to remove the old syntax
highlighter.

Github: #56 “vim highlighting too slow”