add a virtual terminal emulator
This is sort of an admission of failure of the current approach in
`clink_vim_read`. While it is somewhat working, it is a game of whackamole with
handling escape sequences, and the design has them all handled inline. This does
not scale; the core loop is turning into an unmanageable mess of fused
operations and intermingled state updates.

The present commit takes a step back and implements an abstraction for in-memory
terminal emulation that handles a superset of the escape sequences in
`clink_vim_read`. The design can be simpler because we do not need to juggle (1)
an on-the-fly line we are constructing to yield and (2) sequences that
manipulate both the current line and future lines. By more faithfully modelling
a terminal, we can simply apply the effect of an escape sequence and postpone
all reading until the end.

This is not fully plumbed through to `vim-cat`, but I suspect this alternative
design will turn out to be slower, primarily for two reasons:

  1. We will now be constructing the entire rendered file in memory before
     yielding any lines.

  2. Character data is no longer simple `char` strings, but rather “cells” of
     1-4 byte UTF-8 characters.

However, given the fused loop was a ~99.2% speed up over the prior approach
(toHTML.vim), I am hoping this third alternative will retain a reasonable amount
of the speed up while significantly increasing maintainability.