fix: write a full ANSI colour directive when changing cell styles
Previously the emulator would write a directive for each difference. For
example, switching from default style to bold and green would be emitted as
"\033[32m\033[1m". This is fine for regular terminal emulators to consume, but
the translation to Ncurses in clink/src/colour.c assumes each directive will
name an exact colour to switch to. I.e. this directive needs to be emitted as a
single entity, "\033[32;1m".