do Vim renders in blocks of 999 lines
In testing the prior changes, I stumbled across the fact that Vim has a hard
coded limit of 1000 lines (including the `statusline`). Tell it your terminal is
larger than this and too bad, you will get capped at 1000.

This seems like a deal breaker for one-shot rendering. The only work around that
comes to mind is more actively puppeting Vim and scrolling the buffer _as_ we
are reading from it. Even from the idea of this, you can guess this will be
extremely brittle.

So this commit implements a fallback strategy. We render the file in blocks of
999 lines, scrolling to the start of each block before reading so we get the
correct offset. This makes it clear that the prior ~99.2% speed up was because
we were accidentally only rendering 999 lines of the massive 826413 line test
case. Now that we need to render in a loop this may actually end up _slower_
than the previous toHTML.vim strategy.

I am not entirely sure how to judge this. It is such an obvious win for smaller
files (no temporary file creation, no regex and HTML involved, direct path to
rendering, reasonable room for optimisation) and these pathological files I have
been benchmarking are such extreme outliers. For context, none of the other
files in a few moderately sized code bases I have on hand exceed 10000 lines.
While 10000 line sources are not unheard of, even these render at reasonable
speed. It is only once you start getting into the 100000 line range that things
start lagging. And I am not sure this is a reasonable thing to be bothered by;
other tools also struggle with these sources, e.g. MSVC used to (still does?)
OOM on this ~800K source.

In conclusion, lets call this good enough for now and see how things progress.