only send '+call cursor(…)' and '+z' when required
It seems sending these sequences for a file that fully fits in the virtual
screen results in unexpected Vim behaviour. I can reproduce this interactively
– Vim does some sort of paging thing where you get “More” and need to hit space
to continue – but I do not know what it is trying to indicate. There is no
actual message or other information rendered.

Rather than trying to delve into what exactly is going on here, this commit
simply avoids getting into this problematic situation.