do parse and highlight in a single pass
The previous database construction logic looked for files to parse and, as it
saw each such file, it would enqueue it for highlighting. The thinking behind
this design is that a file could be parsed on one thread while being highlighted
on another. In practice, the overhead of work queue synchronisation and an
unnecessary additional push-pop to the work queue makes this not worthwhile.

This commit takes a different strategy and simply does the highlighting
immediately following parsing on the same thread that did the parsing.