scan directories eagerly, push files upfront
This change populates the pending queue immediately when files and directories
are pushed to it. The result is longer time upfront, but then simpler logic for
pop. In particular, one of the benefits of this change is that the number of
pending files is known in advance. We are now in a position to start working on
a progress output for parsing.

Github: #62 “progress output”