drop pushed tracking in the file queue implementation
The file queue attempted to catch and drop the pushing of files that had already
been added to the queue. But this is already done by the inner string queue. A
side effect of switching to pushing everything upfront was that we now only
guard against duplicate file pushes, not duplicate directory pushes, making the
guard redundant in like of the equivalent string queue mechanism.

Allowing duplicate directory pushes seems acceptable. If you pass duplicate or
overlapping paths to Clink, it should be pretty obvious to you and the mistake
is on you. Clink will suppress duplicate file parsing to help you out, but also
attempting to suppress duplicate directory scanning to cope with your mistake
seems like an unnecessary overcompensation.

This holds true in the face of symlinks too: if there is an unknown (to you)
symlink in one of your directories that causes duplication like this, Clink will
help you out by suppressing double-work during parsing. But you will still get a
duplicate directory scan. So be it.