use set-resident strings instead of allocating separate queue entries
This takes advantage of some of the previous changes and reduces dynamic
allocations and associated heap pressure by having the string queue return
pointers into its seen set. Because this set is live for the lifetime of the
queue, we basically get string interning and memory management for free.

This also fixes a latent bug where the path leading to the “failed to make %s
relative” error performed a use-after-free.