restructure 'str_queue' representation
This tweaks not the underlying data layout itself but rather how we represent
the queue head and queue tail. This is all framed around a new introduced
assumption, that all pushes will precede all pops.

This all seems a bit mysterious if you do not know where we are going with this.
The movements of this commit are getting all our ducks in a row for a single RMW
in pop. This will enable us to make pop thread safe while not giving up lock
freedom.

Note that the above goal is only achievable because we scan all paths eagerly
upfront, meaning no push operations happen in a multithreaded context any more.