remove work queue
Clink constructed a work queue that contained a file queue that contained a
string queue. This design formerly made sense as each layer of the onion had its
purpose:

  1. String queue: storing strings, de-duping.
  2. File queue: distinguishing files vs directories, recursing.
  3. Work queue: concurrency, locking.

However after moving to eagerly scanning paths, all pushing was being done in a
single-threaded context. From here it is not a great step to realise pop is the
only operation that needs to be thread safe. A follow on insight from that was
that it was not hard to make the string queue’s pop thread safe. Having done
that, the motivation for layer 3 evaporates. We do not need a concurrency layer
when the lowest layer is already thread safe.

As a bonus, thread safety for (1) was implemented lock-free so we dropped all
mutex usage.

This ended up being a very nice complexity, performance, and maintainability win
all at once.