remove 'parse-c' test binary
This utility was previously added to interactively test the C parsing logic. The
C parser is about to become more complicated and understand compilation
databases. It does not seem worth introducing this complexity into `parse-c` as
opposed to using the `DEBUG` macro now for debugging. Removing this will also
enable us to simplify the C parser in upcoming commits.

Github: related to #46 “cmake mode”