switch to using the generic parser for C/C++
This is a sort of going backwards to go forwards move. The long term direction
we would like to take is calling `clangd` to parse C/C++. Except that `clangd`
needs a compilation database, which users sometimes do not have. So we need a
fallback strategy for parsing C/C++, and this is it.

This parsing is very unsophisticated and could easily be improved in future. But
it is not clear it is worth it, when the real answer to accurate results is to
supply a compilation database and thereby use `clangd`.

Note that this leaves the `argc`, `argv` options to `clink_parse_c` unused.
These may be removed in future.