hand roll a replacement C parser
This is pretty weird thing to do, but the leading comment explains my thinking.
Surprisingly a cursory test drive of this on some open source code bases
indicates it is already reasonably fast and accurate.

Some open TODOs:

  * recognise function calls
  * recognise and ignore string/character literals
  * handle Windows line endings
  * recognise #includes

The last is probably best done with some sort of modular C pre-processor
scanner. Because we will want to reuse this for C++ parsing.

The code in this commit can probably be significantly accelerated as well, but
there is not much point until we get it complete.