more nuanced scope concept in the C parser
This teaches the C parser the notion of '('-based scopes in addition to
'{'-based scopes. I realised we need something like this to capture that
function parameters have a parent of the function itself.

Note that a parent is registered when it is first seen, but then only becomes
active once we enter a scope. This asymmetry avoids some lookahead complications
when seeing a function definition and trying to discriminate whether it is a
full definition or just a prototype.