recognise C functions not immediately preceded by a type
When something has been typedefed or a type we do not immediately recognise as
such is returned from a function, the parser would not recognise this as a
function definition:

  my_type_t foo(int x) {
    // foo was not recognised as a function definition, and thus parent
  }

This commit tweaks how we recognise both function definitions and function calls
to relax both. Like everything else, this is a heuristic, but it seems to work
better than the previous logic.