use span interface for symbol addition during C parsing
It is not clear whether this is worth it. Despite obviously resulting in less
heap allocation, in both debug and release builds this actually _increases_ the
runtime of parsing an example (Clink’s own code base) by ~12%. The culprit
appears to be the more expensive calling convention for `add_symbol` vs
`clink_db_add_symbol`; LTO brings both the before and after states to roughly
the same performance, beating all non-LTO options in this matrix.