C parser: process words while at their end, rather than character after
This change makes the parser somewhat simpler by handling all word processing
while we are still within the word. That is, a pass through the outermost loop
is only _either_ ending a word or doing other generalised processing. Previously
doing the two in a single iteration led to some harder to understand flow.