C parser: ignore qualifiers
Something that should have been obvious to me before was that for our purposes
type qualifiers are completely irrelevant. Note that this change represents a
divergence in that some GCC extensions are accepted. We will likely introduce
even more of this in future.

This leaves a gap in that some qualifiers are not recognised by the parser as
keywords, but the control flow for now makes this benign.