fix: stop ignoring `-I`
Confusingly the handling of this seems to have always been wrong. Include
directories were computed but then discarded. This problem was detected by ASan
in CI after an upcoming change. This probably does not matter much as I was
intending to remove these options, but it is surprising they apparently never
had any effect.