parse: [nfc] rearrange when parent is set
The motivation for this is to allow the logic that decides the category of the
symbol we are about to add to also potentially activate a parent. If retrieving
the parent remained at the bottom of the block, it would make this retrieval
pick up the just-made-active new parent. That is, a symbol would become the
parent of itself.