integrate CPP support into the C parser
The C parser now switches into a C pre-processor mode when seeing a CPP
directive. This CPP parser is so far only capable of recognising `#define` and
then symbol references. The parsers work like a stack, so when the CPP parser
decides it is done (EOL), it pops back out to the calling C parser.