fix writing new records with 0 size and hash
When parsing a file for the first time, it would have no record in the database.
This would result in `stat`ing of the file being skipped entirely. The hash and
timestamp would remain 0 and these values would be inserted into the database.
On a second run, Clink would find the previous record but see the hash and
timestamp as differing from the properties of the file on disk and incorrectly
decide it needed to reparse the file.

Github: fixes #87 “cache logic does not work”