support on-demand syntax highlighting
The ncurses UI can now recognise a symbol missing context (syntax-highlighted
line content) and then perform the highlighting at the time it is constructing
the list of results. There is no support yet for delaying highlighting (avoiding
doing it during the build phase), but this is upcoming.

This implementation is slightly suboptimal in that it will repeatedly try to
highlight the same file if multiple matches to it occur within the same result
set.