reestablish libclang dependency
This is part of an arc back towards a libclang-based C/C++ parser. We have flip
flopped on this a number of times. The crux of the indecision basically boils
down to this: Libclang-based parsing seems to be both slow and unusably
inaccurate when not given precise command line flags (include paths, hash
defines, etc.). The LLVM stance (AFAIK) is that this is libclang “behaving as
expected.” Parsing C++ is quite literally undecidable without precise flags.
Except this is _not_ true for C. And, moreover, a primary use case for Clink is
“I do not know what flags this project is built with, please do your best and
try to parse its source anyway.”

So hand rolling a less persnickety parser seems like a candidate. Except, as
anyone who has tried it will tell you, building your own (even fuzzy) parser for
modern C++ is a non-trivial exercise. The fuzzy C parser in the current tree
does a surprisingly good job. But I can already see that every further
improvement to it will be an exercise in surgical precision, further degrading
its readability. And making it work for C++ as well magnifies this effect
greatly. Cscope’s parser is close to optimal, in the sense that reaching parity
with it let alone improving it requires a _lot_ of engineering effort. None of
this is surprising, nor would it have surprised the me who began this “roll your
own parser” exercise. I just wanted to game this out a little more to confirm my
intuition.

So we are left with the decision: do we want to write our own compiler or do we
want to write our own Cscope? I think it is time for Clink to exit the compiler
business and return to where its strengths lie. On a well-defined code base,
nothing hand rolled will come close to libclang in precision, so lets revert to
using it. If it is slow, lets find out why. If it is inaccurate, lets work with
upstream to make it better.

Along this journey, I also explored the Language Server Protocol (LSP), for
which `clangd` is the witness in the LLVM ecosystem. However I could not
understand how to match Clink with LSP’s API. LSP wants you to tell it things
like “I am at line 10 column 2 in foo.c, what are the code completions of
prin__?” But Clink wants to say “globally, what are all the instances of
printf?” The Workspace part of the API looks like it can give you this, but it
seems either unimplemented or unusable in `clangd`. All my attempts to access it
returned no results, even on trivial projects.

This partially reverts commit badf0f6fcdaddc746ac721428ed4b3c9e80166d4 and fully
reverts 00532e1c027200ac950f3aca667209ad13967036.