libclang-ls: tokenise macro expansions
I think this is the most feasible way to retrieve the contents of parameters to
macro calls, as these otherwise do not appear in the AST at all. This is a bit
unfortunate, as we lose any semantic information about e.g. whether an
identifier was a function call. But it seems the best we can do.