add an abstraction over libclang’s compile commands API
I had been planning to add this into libclink. However I realised that it would
be little more than the libclang API as it was, with some locking added. So it
seems more appropriate to just take a direct libclang dependency in clink itself
and just call the libclang API from there.

Github: #46 “cmake mode”