extract compile_commands.json entries without their source path
When passing full compilation arguments to libclink’s `clink_parse_with_clang`,
libclang rejects calls when it is given _both_ compilation arguments and a
source path. We can suppress the source path when we have arguments, but then
what do we do in the case when we want to pass a source path and default `-I…`
arguments?

Thankfully it seems like CMake only generates compile commands with the filename
as the last argument. And both Clink and CMake treat all paths as absolute in
this area. So we can simply recognise and strip the source path as the last
argument.

All a bit circuitous, but it gets the job done.