insert system includes prior to compilation database command
This inserts the `-isystem` parameters prior to the compilation command from the
database. This solves a problem where the compiler knows how to find parts of
its sysroot so CMake (correctly) does not emit explicit include directives when
writing compilation commands. But Libclang does not know the compiler’s sysroot,
so it is unable to find system includes without further help.

Github: closes #96 “macOS: stdbool.h not found”