postpone macro expansion adding until the end of parsing
This has no immediate effect other than shifting the order in which we add
database entries. The intention here is to prepare to be able to add macro
expansions at the point at which we discover their (lexical) parent. This will
allow us to get proper parent information for macros.