reparent orphaned macro expansions
When encountering something that can serve as a parent (e.g. a function
definition), we now scan the pending macro expansions and see if any of them lie
in the extent of this parent. If they do, then we can infer this is their
parent.

The effect of this is to make function call symbols corresponding to macro
expansions more accurate. Where previously the majority of these would come with
no parent because they were encountered in the initial C pre-processing phase,
they now come parented when they occur within a function (the common case). This
seems sort of trivial internally, but has a big impact on aligning with the
user’s expectations.