port from Ncurses to something internal
The motivation for this was to get simpler colour rendering and a simpler path
to UTF-8 support. But it turns out this dramatically simplifies the code, fixes
some visual artifacts, and increases responsiveness.

Note that the strategy taken by the internal replacement is a little odd in that
we disable echo and never re-enable it. It is simpler to mimic the effect of
echoing by our own manipulation of the screen, ensuring we are always in full
control of the content.