fix off-by-one in adjusting right text in ncurses UI
This is a pretty classic – yet subtle – off-by-one error. The two character
arrays, `left` and `right`, contain the text to the left and right,
respectively, of the cursor during “input” mode. They are typically larger than
their contained text, but the valid range of bytes to access in each is up to
`strlen(foo) + 1`. So an expression like:

  memmove(bar, right + 1, strlen(right) + 1);

actually reads one byte _beyond_ the allowable range of `right`.

The subtlety is how this mistake occurred. This code is trying to account for
the NUL terminator. I.e. it would normally require `+ 1` when copying _from_
`right`. So the expression when copying the entire string from `right + 1`
should be `strlen(right) + - 1 + 1`.

This cannot be caught by ASan and friends because of the underlying
over-allocation of both `left` and `right`. Perhaps we should poison their
unaddressable range in future. Or more likely, simply allocate them tighter and
rely on the underlying malloc to optimize its choice of how to back these
allocations.