support UTF-8 at the input prompt
It is now possible to use the full spectrum of UTF-8 when inputting text to
search for. This is not particularly relevant for symbols. But this is
especially valuable when searching for header includes. Prior to this commit,
there was no way to search for inclusion of a header whose name contained a
non-ASCII UTF-8 character.

For completeness, UTF-8 is fully supported in all fields. This may come in handy
if we start supporting non-C/C++ languages that permit non-ASCII characters in
symbol names.