leave spinner active for entirety of 'format_results'
This leads to even more natural behaviour above and beyond the previous tweak.
The spinner now appears completely fluid the entire time. Note that we use
stdout `flockfile` to ensure the threads do not interfere with the placement of
each other’s output. They can race to `fflush`, but this is benign.