backport 'clink-repl' to Python <3.9
I thought we could get away with requiring Python Python ≥3.9. But because we
do not explicitly check for it anywhere, environments with an earlier version of
Python get cryptic errors when jumping into Vim from Clink. It is simpler to
just properly support Python all the way back to Python 3.7.