eagerly left-trim context lines during database insertion
The “context” for each source file line is stored in the database; a
syntax-highlighted version of the literal text of that line. When it comes time
to display a line in the curses UI, white space is stripped from the left side.
This leads to more natural and less wasteful output.

While the stripping is done at UI formatting time, this is somewhat wasteful:

  1. The on-disk database contains white space characters that are
     doomed-to-be-stripped bloating its size.¹
  2. The cost of stripping the string is paid repeatedly, once each time we
     receive a result including the given context line.

This commit alters the stripping to be done at database insertion time. This
means the database size is reduced (point 1 above) and stripping is done
precisely once (point 2 above).

The UI rendering stripping is still done, but this could be dropped in future.

Github: closes #133 “ltrim during database construction”

¹ This is not particularly significant. On Graphviz commit
  15103239103e2481d2f7c1d1bf1eeb13c2c3bd6d this costs ~673KB. Still, why pay
  any overhead if we do not have to?