fix: force clang parser in the test suite
The change in 3fec779f6a9a18daa3a2661abc204704fe29ab8b made these tests default
to Cscope when it was installed. This continued passing in CI where Cscope is
not installed but would fail locally. These tests are supposed to probe the
Clang-based parser. We care less about the Cscope-based one whose behaviour is
known and not changeable.