pre-compile regexes when running searches
Previously the `regexp` function we were installing into the SQLite database
connection was compiling and destroying its regex every invocation. So for a
symbol search that scanned 1000 rows, this would result in 1000 `regcomp` and
`regfree` calls.

This change makes regexes now compiled upfront when a search is invoked. On a
representative test case, this is a ~88% speed up. This is large enough to have
a significant difference on how responsive Clink feels when interactively using
the UI.

Regexes are maintained in a linked-list and never freed until database close.
Although this means regexes outlive their usage within an individual search,
this seemed like a good compromise. This means `re_find` only needs its first
load (of the head pointer) to be atomic (though typical usage will result in
this function finding what its looking for exactly at the head anyway).